Fix db class with flask
